refactor: app design concept (#305)
