+++ embedchain/config/__init__.py
@@ -1,5 +1,7 @@
 from .AddConfig import AddConfig, ChunkerConfig  # noqa: F401
+from .apps.AppConfig import AppConfig  # noqa: F401
+from .apps.CustomAppConfig import CustomAppConfig  # noqa: F401
+from .apps.OpenSourceAppConfig import OpenSourceAppConfig  # noqa: F401
 from .BaseConfig import BaseConfig  # noqa: F401
 from .ChatConfig import ChatConfig  # noqa: F401
-from .InitConfig import InitConfig  # noqa: F401
 from .QueryConfig import QueryConfig  # noqa: F401
